feature(core): clean code
